refactored sales analyst
