+++ lib/invoice_item.rb
@@ -22,9 +22,6 @@ class InvoiceItem
   end
 
   def unit_price_to_dollars
-    # repeated method, consider moving to module
     unit_price.to_f
   end
-
-
 end
