+++ lib/item.rb
@@ -22,12 +22,10 @@ class Item
   end
 
   def unit_price_to_dollars
-    # repeated method, consider moving to module
     unit_price.to_f
   end
 
   def merchant
     parent.find_merchant(merchant_id)
   end
-
 end
