+++ lib/item_repository.rb
@@ -11,13 +11,13 @@ class ItemRepository
   end
 
   def find_by_id(id)
-    @all.find do |item|
+    all.find do |item|
       id == item.id
     end
   end
 
   def find_by_name(name)
-    @all.find do |item|
+    all.find do |item|
       name == item.name
     end
   end
